### 2024-02-24 #161 @MetinSa
ZodiPy
Submitting Author: Metin San (@metinsa)
All current maintainers: (@metinsa, @dncnwtts, @hermda02)
Package Name: ZodiPy
One-Line Description of Package: Zodiacal emission simulations in timestreams or HEALPix for solar system observers.
Repository Link:  https://github.com/Cosmoglobe/zodipy
Version submitted:   v.0.8.5
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10999612.svg)](https://doi.org/10.5281/zenodo.10999612)
Version accepted: v.0.9.2

---

ZodiPy is a new Python tool for modelling the zodiacal emission seen by an arbitrary Solar System observer, which can be used for removal of both thermal emission and scattered sunlight from interplanetary dust in astrophysical data. One of our main goals with ZodiPy is to make zodiacal emission simulations more accessible by providing the community with a simple Python interface to existing models. We recently published a [paper on ZodiPy](https://arxiv.org/abs/2205.12962) (which has also been accepted to A&A) where we describe in more detail modelling of zodiacal emission and the approach taken by ZodiPy.

ZodiPy uses Astropy both in the public API and internally. When using one of the methods provided to simulate the zodiacal emission, the user needs to use `Quantity` and `Time` types in input arguments such as frequency/wavelength, angles on the sky, and the time of observation. Internally, the Time object is used to compute the heliocentric ecliptic position of the Earth and the observer with the `SkyCoord` and other functionality in `astropy.coordinates`.

ZodiPy was already [proposed and reviewed as an Astropy Affiliated package](https://github.com/astropy/astropy.github.com/pull/495) before the recent partnership between Astropy and pyOpenSci in [APE22](https://github.com/astropy/astropy-APEs/blob/main/APE22.rst#in-a-nutshell), so I am resubmitting the proposal as is here.

- [x] Astropy: Link coming soon to standards

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

      - The target audience is astronomers and cosmologists who work with sky observations all the way from the optical to CMB-frequencies. The scientific application of the this package is to provide the astronomer with and open source tool to create simulated timestreams, snapshots, or maps of the zodiacal light as seen from earth or any other arbitrary solar system observer. This allows existing experiments with data contaminated by zodiacal light to model and remove the light from their data, and future experiments can, and some already has ([Nancy Grace Roman Space Telescope](https://ui.adsabs.harvard.edu/abs/2023arXiv230617226R/abstract)) use ZodiPy to make forecasts and simulations.

     -  There exists a [few zodiacal light tools](https://lambda.gsfc.nasa.gov/product/foreground/fg_models.html#zodiacal_models]) but these are all experiment specific and can only be used to model the zodiacal light that was observed in their data. ZodiPy is as far as i know the only Python package that allows for arbitrary solar system observers to make zodiacal light simulations. ZodiPy is also extendible and can easily be updated with new improvements to zodiacal light models and even used to fit new models.
