### 2024-02-28 #161 @isabelizimm
---
---

Hello there and welcome to the pyOpenSci universe (apologies for the bad pun 😂🌟)! We're happy you're here. I've done a quick overview of the `zodipy` package, and you are in great shape! **There are a few things we will need before we start with your review:**

1. Please add a `CODE_OF_CONDUCT.md` to the root of your repository. For more information, pyOpenSci's packaging guide has some [examples and information on CoC files.](https://www.pyopensci.org/python-package-guide/documentation/repository-files/code-of-conduct-file.html)
2. Please add a `CONTRIBUTING.md` to the root of your repository. For more information, pyOpenSci's packaging guide has some [templates and information on CONTRIBUTING files.](https://www.pyopensci.org/python-package-guide/documentation/repository-files/contributing-file.html)
3. I'll also nudge each maintainer (so, @MetinSa @dncnwtts) to [fill out our onboarding survey](https://forms.gle/F9mou7S3jhe8DMJ16).

Let us know when you've made these updates. I'll also use this message to introduce your editor-- @dhomeier will be helping you all through the review process 🎉 Thanks for taking on this role Derek!
