### 2024-02-28 dd3170f
Add contributing guidelines and a code of conduct
