+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+# Contributing Guidelines
+
+
+Thank you for considering to contribute to ZodiPy!
+This is a community-driven project, so it's people like you that make it useful and
+successful. These are some of the many ways to contribute:
+
+* :bug: Submitting bug reports and feature requests
+* :memo: Writing tutorials or examples
+* :mag: Fixing typos and improving the documentation
+* :bulb: Writing code for everyone to use
+* :people_holding_hands: Community engagement and outreach
+
+If you get stuck at any point you can create an
+[issue](https://github.com/GenericMappingTools/pygmt/issues) on GitHub
+or contact us at one of the channels mentioned
+in the [Contributing Guide](https://www.pygmt.org/dev/contributing.html).
+
+For more information on contributing to open source projects,
+[GitHub's own guide](https://opensource.guide/how-to-contribute)
+is a great starting point if you are new to version control. Also, checkout the
+[Zen of Scientific Software Maintenance](https://jrleeman.github.io/ScientificSoftwareMaintenance/)
+for some guiding principles on how to create high quality scientific software
+contributions.
+
+
+## Ground Rules
+
+The goal is to maintain a diverse community that's pleasant for everyone.
+**Please be considerate and respectful of others**. Everyone must abide by our
+[Code of Conduct](https://github.com/Cosmoglobe/zodipy/blob/main/CODE_OF_CONDUCT.md)
+and we encourage all to read it carefully.
