### 2024-02-28 246c535
Fix link in contributing guidelines
