### 2024-02-28 #161 @MetinSa
Hi! :wave:

I have now added a `CODE_OF_CONDUCT.md` and a `CONTRIBUTING.md` to the repository aswell as having filled out the survey. I also added @hermda02 as a maintainer of the project.
