### 2024-03-20 Cosmoglobe/zodipy#3 @pllim
parallel value can be inferred from n_proc (no need for parallel)
In `Zodipy` (https://cosmoglobe.github.io/zodipy/reference/), `n_proc=1` can be inferred as `parallel=False`, otherwise `parallel=True`. Making user specify both `parallel` _and_ `n_proc` seems unnecessary. Is there a reason why they have to be separate?

**Describe the solution you'd like**

`parallel` keyword can probably be deprecated and then removed without affecting user-facing functionality.

N/A

**Additional context**

This is a part of:

* https://github.com/pyOpenSci/software-submission/issues/161
