### 2024-03-20 Cosmoglobe/zodipy#4 @pllim
DOC: No dev install instructions or mention of non-standard dependencies in README
Missing: the development version of the package and any non-standard dependencies in README.

This is a part of:

* https://github.com/pyOpenSci/software-submission/issues/161
