### 2024-03-20 Cosmoglobe/zodipy#5 @pllim
DOC: Missing badge from gh-pages deployment workflow
Missing: badge for docs building. Could add a badge from https://github.com/Cosmoglobe/zodipy/actions/workflows/pages/pages-build-deployment (assuming that is the workflow that would be green if cosmoglobe.github.io/zodipy/ is deployed successfully).

This is a part of:

* https://github.com/pyOpenSci/software-submission/issues/161
