### 2024-03-20 Cosmoglobe/zodipy#6 @pllim
DOC: Missing badge for 
Missing: badge for [repostatus.org](https://www.repostatus.org/).

This is a part of:

* https://github.com/pyOpenSci/software-submission/issues/161
