### 2024-03-20 Cosmoglobe/zodipy#7 @pllim
DOC: missing badge for Python versions supported
Missing: badge for Python versions supported.

This is a part of:

* https://github.com/pyOpenSci/software-submission/issues/161
