### 2024-03-21 795e837
Remove `parallel` keyword, and make parallelization be determined by `n_proc` which now defaults to 1, in which case we have no parallelization.
