### 2024-03-21 9e8c211
Add test for multiprocessing of model with components featuring inner radial cutoffs
