### 2024-03-21 e88fefc
Update pyproject for new release and add classifiers
