### 2024-03-21 53afb08
Add python 3.12 version classifier since this was not auto detected by poetry/pypi, and add python 3.12 to test workflow
