### 2024-03-21 21dd61a
Add badges. Closes #5, closes #6, and closes #7.
