+++ README.md
@@ -23,6 +23,14 @@ See the [documentation](https://cosmoglobe.github.io/zodipy/) for more informati
 # Installation
 ZodiPy is installed using `pip install zodipy`.
 
+# Dependencies
+ZodiPy supports all Python versions >= 3.8, and has the following dependencies:
+- [Astropy](https://www.astropy.org/) (>=5.0.1)
+- [NumPy](https://numpy.org/)
+- [healpy](https://healpy.readthedocs.io/en/latest/)
+- [jplephem](https://pypi.org/project/jplephem/)
+
+
 # A simple example
 ```python
 import astropy.units as u
@@ -44,6 +52,53 @@ emission = model.get_emission_ang(
 print(emission)
 #> [15.35392831 15.35495051 15.35616009] MJy / sr
 ```
+
+# For developers
+Contributing developers will need the download the following additional dependencies:
+- pytest
+- pytest-cov
+- hypothesis
+- coverage
+- ruff
+- mypy
+- pre-commit
+- mkdocs
+- pymdown-extensions
+- markdown-include
+- mkdocs-material
+- mkdocstrings
+- mkdocstrings-python
+- markdown (<3.4.0)
+
+which are required to test and build ZodiPy.
+
+## Poetry
+Developers can install ZodiPy through [Poetry](https://python-poetry.org/) (Poetry >= 1.8.0) by first cloning or forking the repository, and then running 
+```
+poetry install
+```
+in a virtual environment from the repository root. This will read the `pyproject.toml` file in the repository and install all dependencies. 
+
+## pip
+Developers not using Poetry can install ZodiPy in a virtual environment with all dependencies with pip by running
+```
+pip install -r requirements-dev.txt
+```
+from the repositry root. This will read and download all the dependencies from the `requirements-dev.txt` file in the repositry. 
+
+Note that developers using Python 3.12 will need to upgrade their pip versions with `python3 -m pip install --upgrade pip` before being able to install ZodiPy. This is due to known incompatibilities between older pip versions and Python 3.12
+
+## Tests and formatting
+The following tools should be run from the root of the repository and producing no errors. These are ran automatically as part of the CI workflows on GitHub, but should be tested locally first
+
+- [pytest](https://docs.pytest.org/en/8.0.x/): Tests are run with pytest by simply running `pytest` in the command line in the repository root. 
+- [ruff](https://github.com/astral-sh/ruff): Formating linting is done with `ruff` by simply running `ruff` in the command line in the repository root. 
+- [mypy](https://mypy-lang.org/): Type checking is done with `mypy` by simply running `mypy zodipy` in the root of the repository.
+
+Remeber to add tests when implementing new features to maintain a high code coverage.
+
+## Documentation
+We use [MkDocs](https://www.mkdocs.org/) to create our documentation. Documentation is automatically built and uploaded as part of the github CI. New features should be documented by adding content to the `docs/` folder at the appropriate places.
 
 # Scientific paper and citation
 For an overview of the ZodiPy model approach and other information regarding zodiacal emission and interplanetary dust modeling we refer to the scientific paper on ZodiPy:
