### 2024-03-21 Cosmoglobe/zodipy#10 @pllim
poetry install downgrades numpy and fails
**Describe the bug**

Trying to install using `poetry install` as documented but it fails with:

```
- Downgrading numpy (1.26.4 -> 1.24.4): Failed
```

**Expected behavior**

Just use installed numpy if compatible. Any reason why this is incompatible with numpy 1.26? There is no mention of such incompatibility in README.

 - OS: Ubuntu (WSL2 on Windows 11)
 - Browser [e.g. chrome, safari] N/A
 - Version [e.g. 22] N/A

**Additional context**

This is part of:

* https://github.com/pyOpenSci/software-submission/issues/161
