### 2024-03-21 #161 @pllim
- [x] **Installation instructions:** for the development version of the package and any non-standard dependencies in README. (Reviewer note: https://github.com/Cosmoglobe/zodipy/issues/4 and https://github.com/Cosmoglobe/zodipy/issues/10)

    - [x] Docs building (if you have a documentation website) (Reviewer note: https://github.com/Cosmoglobe/zodipy/issues/5),
    - [x] A [repostatus.org](https://www.repostatus.org/) badge (Reviewer note: https://github.com/Cosmoglobe/zodipy/issues/6),
    - [x] Python versions supported (Reviewer note: https://github.com/Cosmoglobe/zodipy/issues/7),

- [ ] ~Any additional setup required to use the package (authentication tokens, etc.)~ (Reviewer note: N/A)
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem. (Reviewer note: I do not know enough about zodiacal light community to check off this one.)

- [x] **Packaging guidelines**: The package conforms to the pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide). (Reviewer note: Not using src layout but should not matter.)

---

#### Astropy-specific

- [x] **Relevance to astronomy/astrophysics:**
    - [ ] **Out of scope:** Not useful for astronomers, or specific to one project/collaboration. (This is a basis for rejection.)
    - [x] **Specialized package:** Useful to astronomers working in a very specific domain/field, or with a specific telescope instrument and usable not just by a single collaboration but any astronomers within that domain. Packages such as sncosmo fall into this category.
    - [ ] **General package:** Package that is useful for astronomers across more than a single field/instrument/telescope. Packages such as astroquery or astroplan fall into this category.
- [x] **Integration with Astropy ecosystem:**
    - [ ] **No integration:** Does not use Astropy or other Astropy Affiliated packages anywhere where it should be possible, and/or uses other libraries instead, or unnecessarily duplicates functionality found in Astropy or other Astropy Affiliated packages. (This is a basis for rejection.)
    - [x] **Partial integration:** Makes an effort to use Astropy or other Astropy Affiliated packages in places, but still has other places where this could be done.
    - [ ] **Good integration:** Uses Astropy or other Astropy Affiliated packages wherever possible. Where not, there are good reasons not to.
For specific actionable items, reviewer or editor are free to open issues on the package repository directly and link them

I skipped the JOSS section because I don't see a `paper.md` and so assume this package isn't asking for JOSS listing also.

This package is already in Astropy's pre-[APE 22](https://github.com/astropy/astropy-APEs/blob/main/APE22.rst) legacy listing over at https://www.astropy.org/affiliated/legacy.html via https://github.com/astropy/astropy.github.com/pull/495 .
