### 2024-03-22 #161 @MetinSa
I'm sorry, I had forgotten about the JOSS section but would very much like a JOSS listing. I will start working on a `paper.md`.
