### 2024-03-22 f5f723b
Fixes #12 where hypothesis `lists` strategy no longer worked after updating hypothesis version.
