### 2024-03-22 Cosmoglobe/zodipy#10 @MetinSa
If you have time, could you check if this works as intended on the `drop_py38` branch? If so, is it sufficient to just release a >=3.9 version as a fix to this?
