### 2024-03-22 c665d5f
Fix missing caret in hypothesis version in `pyproject.toml` [no ci]
