### 2024-03-22 Cosmoglobe/zodipy#10 @pllim
Seems to work on `drop_py38` but only after I run `poetry lock` also. Otherwise, I get this:

```
$ poetry install
Installing dependencies from lock file

pyproject.toml changed significantly since poetry.lock was last generated. Run `poetry lock [--no-update]` to fix the lock file.
```

Hope this helps!
