### 2024-04-10 20a7597
Update dependencies (SciPy was not listed)
