### 2024-04-10 f5afb46
Add github action for JOSS pdf generation
