+++ docs/install.md
@@ -12,8 +12,8 @@ pip install zodipy
 ## Dependencies
 ZodiPy has the following dependencies (these are automatically downloaded alongside ZodiPy):
 
-- [numpy](https://numpy.org)
-- [astropy](https://www.astropy.org) (>= 5.0.1)
+- [Astropy](https://www.astropy.org) (>= 5.0.1)
+- [NumPy](https://numpy.org)
 - [healpy](https://healpy.readthedocs.io/en/latest/)
 - [jplehem](https://pypi.org/project/jplephem/)
-- [scipy](https://scipy.org/)
+- [SciPy](https://scipy.org/)
