### 2024-04-10 2933ab4
Add possibility to run pdf compilation on workflow dispatch [no ci]
