+++ paper/paper.md
@@ -46,8 +46,8 @@ light tools (see the
 [LAMBDA foreground models page](https://lambda.gsfc.nasa.gov/product/foreground/fg_models.html) 
 for a list of existing tools). However, these programs are either 
 only usable for specific experiments or otherwise difficult to access by requiring 
-licensed programming languages or web  interfaces. Many modern astronomy and 
-cosmology pipelines are built in Python due to the wide range of available 
+licensed programming languages or the use of web interfaces. Modern astronomy and 
+cosmology pipelines are commonly built in Python due to the wide range of available 
 high-quality tools and open-source projects and communities, such as the Astropy 
 project [@astropy]. The lack of a general-purpose zodiacal light tool in this space 
 was the primary motivation behind the development of the `ZodiPy` package.
@@ -59,9 +59,9 @@ scratch. To use `ZodiPy`, the user is required to provide the following data:
 1) A sequence of pointings, either in ecliptic or galactic coordinates. These can 
 be specified  as angles on the sky or as HEALPix [@Gorski2005] pixel indices.
 2) A center frequency or an instrument bandpass.
-3) The time of observation, which corresponds to the pointing sequence. 
+3) Time of observation, corresponding to the pointing sequence. 
 4) The heliocentric ecliptic position of the observer. If the observer is located 
-at a major solar system object, such as the Earth or the Sun-Earth-Moon barycenter 
+at a major Solar system object, such as the Earth or the Sun-Earth-Moon barycenter 
 L2, the position is instead queried through the `astropy.coordinates` Solar system 
 ephemerides. 
 
@@ -69,20 +69,20 @@ The predicted zodiacal light is then obtained by evaluating a sequence of
 line-of-sight integrals from the position of the observer and through a 
 model of the three-dimensional interplanetary dust distribution. For 
 implementation details and examples of how to apply `ZodiPy` to a real-world 
-dataset, see[@San2022].
+dataset, see @San2022.
 
 `ZodiPy` has been rapidly adopted by the astronomy community, and the package has 
-already been used by several research projects[@2023arXiv230617219A; 
-@2023arXiv230617226R; @Tsumura2023; @Avitan2023; @Hanzawa2024] with applications 
-ranging from assisting the coming NASA Roman Space Telescope in determining its 
-observational fields to modeling data obtained aboard the Hayabusa2 JAXA satellite. 
+already been used by several research projects (see @2023arXiv230617219A; 
+@2023arXiv230617226R; @Tsumura2023; @Avitan2023; @Hanzawa2024), for instance in
+assisting the coming NASA Roman Space Telescope in determining its observational 
+fields, or to modeling data obtained aboard the Hayabusa2 JAXA satellite. 
 
 # Acknowledgements
-The work to develop `ZodiPy`  has received funding from the European Union’s Horizon 
-2020 research and innovation programme under grant agreement numbers 819478 
-(ERC; Cosmoglobe) and 772253 (ERC;bits2cosmology).
+This work has received funding from the European Union’s Horizon 2020 research and 
+innovation programme under grant agreement numbers 819478 (ERC; Cosmoglobe) and 
+772253 (ERC;bits2cosmology).
 
-This project relies on the following Python packages: Astropy [@astropy], NumPy 
+This project uses the following Python packages: Astropy [@astropy], NumPy 
 [@numpy2011; @numpy2020], healpy [@Zonca2019], SciPy [@scipy2020], and jplephem 
 [@2011ascl.soft12014R].
 
