### 2024-04-10 #161 @MetinSa
Apologies for the delay. I have now added a paper directory containing the `paper.md` file for the JOSS listing.
