+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "zodipy"
 homepage = "https://github.com/Cosmoglobe/zodipy"
-version = "0.8.7"
+version = "0.9.0"
 description = "Software for simulating zodiacal emission"
 authors = ["Metin San <metinisan@gmail.com>"]
 readme = "README.md"
@@ -30,6 +30,7 @@ numpy = "^1.26.4"
 healpy = "^1.16.6"
 astropy = ">=5.0.1"
 jplephem = "^2.17"
+scipy = "^1.13.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
@@ -43,9 +44,9 @@ pytest-cov = "^4.0.0"
 mkdocs-material = "^9.0.1"
 mkdocstrings = "^0.23.0"
 mkdocstrings-python = "^1.7.3"
-ruff = "^0.1.4"
+ruff = "^0.3.7"
 markdown = "<3.4.0"
-hypothesis = "6.99.11"
+hypothesis = "^6.99.11"
 
 
 [build-system]
@@ -69,7 +70,7 @@ filterwarnings = [
 ]
 
 [tool.ruff]
-select = [
+lint.select = [
     "F", 
     "D", 
     "F", 
@@ -100,7 +101,7 @@ select = [
     "RUF",
 ]
 line-length = 100
-ignore = [
+lint.ignore = [
     "B905",
     "D100",
     "D104",
@@ -111,6 +112,4 @@ ignore = [
     "ISC001"
 ]
 exclude = ["tests/*", "docs/*"]
-
-[tool.ruff.pydocstyle]
-convention = "google"
+lint.pydocstyle.convention = "google"
