### 2024-04-13 Cosmoglobe/zodipy#10 @MetinSa
@pllim hopefully this is resolved in the latest merge where python 3.8 is dropped and the `pyproject.toml`and the lockfile are updated, but feel free reopen this issue if you notice any issues!
