### 2024-04-16 #161 @pllim
FYI -- I have updated the checkboxes for JOSS at https://github.com/pyOpenSci/software-submission/issues/161 for my part of the review.

p.s. I never submitted anything to JOSS before so I cannot answer the questions about JOSS timeline. Hopefully others can. Thanks for your patience!
