### 2024-04-17 #161 @lwasser
hi 👋 i can help here. it's normally quick IF the submittor makes it clear when the package is submitted that:

1. The package has already been reviewed and accepted by pyOpenSci
2. Links to this review here to make it easier for the editor to see this is a fast track
3. The editor for this review reminds the JOSS editor that this is a fast tracked submission (which means that no code needs to be re-reviewed). All they have to do is a) check for scope and b) review your paper.

I've seen this go wrong when there is confusion regarding the fast track part and then things take a bit longer. but otherwise it should be a few weeks if the JOSS editor follows the process and the pyOS editor kind of keeps an eye on things !  i hope that helps.
