### 2024-04-18 #161 @lpsinger
Estimated hours spent reviewing: 2 (includes original review as an Astropy affiliated package)

---

Mostly reproduced from my Astropy affiliated package review.

ZodiPy looks like a very healthy package. Here are some spots where there is room for improvement.

1. It could have much more integration with Astropy core and Astropy affiliated packages. Aside from Astropy units, the public API does not have much integration with Astropy. Here are some issues:

- The get_emission_ang method basically depends on two coordinates that define the ray from the observer to the direction in which they are looking (plus a time and a wavelength). The vertex of the ray can be specified in Cartesian coordinates and the direction of the ray in polar coordinates. Internally, the method instantiates SkyCoord objects in order to do positional astronomy calculations. It would be ideal if the user could specify both coordinates using SkyCoord instances. That would unify 5 arguments (theta, phi, obs, obs_pos, lonlat) into only two.
- Conversion between coordinate frames is done using healpy.Rotator but could be done more simply and idiomatically with SkyCoord frame transformations.
- There is a synthetic photometry feature to integrate the observed flux over a bandpass. ZodiPy could employ the Astropy affiliated package synphot here.
- ZodiPy can evaluate the emission on HEALPix maps, but uses healpy to do this. It could use the Astropy affiliated package astropy-healpix here.
- ZodiPy can plot emission components on all-sky maps, but uses healpy to do this. It could use reproject + astropy-healpix here for more flexible plotting.

2. The approaches to Numpy vectorization and Python multiprocessing parallelism are somewhat fixed function, and may not be performant for all use cases.

- Methods do not support Numpy broadcasting for arbitrary input shapes. For example, it appears that obs_time must be scalar. For another, this line in get_obs_and_earth_positions discards the shape of the input arguments:

    return obs_position.reshape(3, 1, 1).value, earth_position.reshape(3, 1, 1).value

- ZodiPy supports parallelism using Python multiprocessing, but it creates its own process pool and discards it when it is done. It would be nice if it supported the pool being passed by the user so that the cost of spawning child Python interpreters is not borne repeatedly. Also there are situations where it is necessary to use a custom Pool subclass (e.g. billiard, https://pypi.org/project/billiard/).

3. Deviations from pyOpenSci [packaging guidelines](https://www.pyopensci.org/python-package-guide):

- Needs a conda-forge package.
- The value of the license in the project metadata (tool.poetry.license) is not an [SPDX license identifier](https://spdx.org/licenses/). Please set it to either of the allowed values, `GPL-3.0-only` or `GPL-3.0-or-later`.
- Consider adding more project URLs (`project.urls` section in pyproject.toml). For project URL names that PyPI displays prominently with icons, see https://github.com/pypi/warehouse/blob/main/warehouse/templates/packaging/detail.html.

4. Code style

The project uses ruff for code formatting, but it is only run on the module itself. Ruff would make changes to the Python modules in the tests directory. Please consider running ruff on the entire project, including the tests.
