### 2024-04-19 44b3cc0
Update ruff configuration so that it runs on all files in the repository.
