+++ tests/test_params.py
@@ -3,16 +3,19 @@ import random
 from hypothesis import given
 
 from zodipy.zodipy import Zodipy
+
 from ._strategies import model
 
 
 @given(model())
-def test_get_parameterse(model: Zodipy) -> None:
+def test_get_parameters(model: Zodipy) -> None:
+    """Tests that the parameters are returned as a dictionary."""
     assert isinstance(model.get_parameters(), dict)
 
 
 @given(model())
-def test_update(model: Zodipy) -> None:
+def test_update_model(model: Zodipy) -> None:
+    """Tests that the model can be updated."""
     parameters = model.get_parameters()
     comp = random.choice(list(parameters["comps"].keys()))
     parameter = random.choice(list(parameters["comps"][comp]))
