+++ pyproject.toml
@@ -5,7 +5,7 @@ version = "0.9.0"
 description = "Software for simulating zodiacal emission"
 authors = ["Metin San <metinisan@gmail.com>"]
 readme = "README.md"
-license = "GPL-3.0"
+license = "GPL-3.0-or-later"
 exclude = ["test.py"]
 classifiers = [
     "Intended Audience :: Science/Research",
