### 2024-04-19 7ae9119
Add urls in pyproject.toml [no ci].
