+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "zodipy"
-version = "0.9.0"
+version = "0.9.1"
 description = "Software for simulating zodiacal emission"
 authors = ["Metin San <metinisan@gmail.com>"]
 readme = "README.md"
