### 2024-04-19 Cosmoglobe/zodipy#19 @MetinSa
Allow users to use astropy `SkyCoord` objects directly in the public API
from https://github.com/pyOpenSci/software-submission/issues/161:

Allowing users to specify the vertex (`obs_pos`) and the direction of the line-of-sight ((`theta`, `phi`) or `pixels`) through an astropy `SkyCoord` object directly in the call to the `get_emission_x` functions. ZodiPy already builds this object internally, so this would provide better integration with Astropy and essentially unify five input arguments (`theta`, `phi`, `obs`, `obs_pos`, `lonlat`).
