### 2024-04-19 Cosmoglobe/zodipy#20 @MetinSa
Use astropy `SkyCoord` to perform rotation between coordinate frames
from https://github.com/pyOpenSci/software-submission/issues/161:

Instead of using the healpy `rotator.Rotator` object, we should use the `SkyCoord` objects directly for coordinate transformation. This would use the already initialized `SkyCoord` objects and integrate better with Astropy.
