### 2024-04-19 Cosmoglobe/zodipy#21 @MetinSa
Use `synphot` package to perform bandpass integration
from https://github.com/pyOpenSci/software-submission/issues/161:

Instead of manually performing bandpass integrations with the custom `Bandpass` type, we should use the Astropy affiliated [`synphot`](https://synphot.readthedocs.io/en/latest/) package. This would also provide better integration with Astropy.
