### 2024-04-19 Cosmoglobe/zodipy#22 @MetinSa
Allow users to optionally pass in multiprocessing pools
from https://github.com/pyOpenSci/software-submission/issues/161:

The user should be able to pass in a multiprocessing pool to ZodiPy, rather than ZodiPy always spawning child Python interpreters (see https://pypi.org/project/billiard/).
