### 2024-04-19 Cosmoglobe/zodipy#23 @MetinSa
Use `astropy-healpix` when projecting to HEALPIx
from https://github.com/pyOpenSci/software-submission/issues/161:

Transforming between angles on the sky and HEALPix pixels can be done with [`astropy-healpix`](https://astropy-healpix.readthedocs.io/en/latest/) instead of `healpy` for better Astropy integration.
