### 2024-04-19 #161 @MetinSa
Thank you for the review @lpsinger!

I have made the following changes after going through the review:
- Ruff now runs on the entire repository (https://github.com/Cosmoglobe/zodipy/commit/44b3cc04ffae136d698e9e4082e7b3597f2bbb59)
- The licence is changed to a SPDX license identifier (https://github.com/Cosmoglobe/zodipy/commit/5eb8df7276188417642fef72ff517b4d7e50f231))
- The `pyproject.toml` file now includes urls to the documentation and repository (https://github.com/Cosmoglobe/zodipy/commit/7ae9119c06088a86e7b9c9143c0934ca25c3908b)
- A PR is open to add ZodiPy as a conda-forge package (https://github.com/conda-forge/staged-recipes/pull/26097)

Additionally, I have opened several [issues](https://github.com/Cosmoglobe/zodipy/issues) (see the mentions above) to address your suggestions for improvement and to better integrate ZodiPy with Astropy. I will start working on these once as soon as I find the time.

Thanks!
