### 2024-04-19 #161 @dhomeier
Thank you for your careful and helpful reviews @pllim and @lpsinger, and for your quick reply @MetinSa!
I'll second the recommendation to include the tests in code style checks, as in a similar attempt in astropy/astropy#16158 we have discovered a number of tests that were not testing exactly the intended functionality.
That should not block acceptance of your package at this time, though.

@pllim I only found one reference, PyEphem, without a DOI in `paper.bib`, but it does not seem to have one in the ASCL entry either – do you think it should have one?
