### 2024-04-19 #161 @MetinSa
I think the reference section is supposed to be empty in the markdown file. If you take a look at the paper artifact produced by the JOSS gh-action here https://github.com/Cosmoglobe/zodipy/actions/runs/8750228773 the references does seem to compile correctly.
