### 2024-04-19 #161 @dhomeier
> I'm not familiar with what JOSS wants in its References section. If it's ok to leave empty, then I'm fine with that. All in all, I don't think it should block acceptance.

The references are all there in the paper.bib file – I was assuming JOSS would accept a standard LaTeX setup, but not actually sure how it handles it with a .md main text – @lwasser do you know if this will work?
