### 2024-04-19 #161 @MetinSa
> I'm not familiar with what JOSS wants in its References section. If it's ok to leave empty, then I'm fine with that. All in all, I don't think it should block acceptance.

The reference section look all right in this [paper draft](https://github.com/Cosmoglobe/zodipy/actions/runs/8750228773) compiled with the JOSS gh-action.
