### 2024-04-19 #161 @dhomeier
Yes, I've checked that off and the reviewers have already checked their final approval.
Just waiting if there is a short term fix for the conda-forge build, but since that is really a healpy issue, it should not block either.
