### 2024-04-19 #161 @pllim
Ah, sorry I misunderstood how JOSS formatting works. I will check off the References box now. Thanks for the clarifications!
