### 2024-04-19 #161 @dhomeier
Great, with conda-forge succeeding for Linux and macOS, this should be ready for JOSS submission.
Congrats @MetinSa! 🎉
