### 2024-04-19 2c03c58
Add pyOpenSci peer-review badge to `README.md` [no ci].
