### 2024-04-19 b5d0031
Add Zenodo doi badge and bump version number [no ci].
