### 2024-04-19 #161 @MetinSa
🎉🎉 I have completed the Zenodo and pyOpenSci peer-review steps(https://github.com/Cosmoglobe/zodipy/commit/2c03c586ae0b4fe7ab06d12d2115c215dbdf9b5d, https://github.com/Cosmoglobe/zodipy/commit/b5d0031fe7fc3d6c05c5aec942c6a067b86c5cdd) and will fill out the post-review survey tomorrow

Thank you to everyone here for the reviews and all the help:)
