### 2024-04-20 #161 @MetinSa
JOSS paper is approved and published https://joss.theoj.org/papers/10.21105/joss.06648!
