### 2024-04-20 3594196
Update `README.md` and `CITATION.bib`with JOSS paper [no ci].
