+++ CITATION.bib
@@ -1,17 +1,30 @@
-@ARTICLE{zodipy,
-       author = {{San}, M. and {Herman}, D. and {Erikstad}, G.~B. and {Galloway}, M. and {Watts}, D.},
-        title = "{COSMOGLOBE: Simulating zodiacal emission with ZodiPy}",
-      journal = {\aap},
-     keywords = {interplanetary medium, cosmic background radiation, zodiacal dust, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2022,
-        month = oct,
-       volume = {666},
-          eid = {A107},
-        pages = {A107},
-          doi = {10.1051/0004-6361/202244133},
-archivePrefix = {arXiv},
-       eprint = {2205.12962},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022A&A...666A.107S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{San2022,
+  author        = {{San}, M. and {Herman}, D. and {Erikstad}, G.~B. and {Galloway}, M. and {Watts}, D.},
+  title         = {{COSMOGLOBE: Simulating zodiacal emission with ZodiPy}},
+  journal       = {\aap},
+  keywords      = {interplanetary medium, cosmic background radiation, zodiacal dust, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+  year          = 2022,
+  month         = oct,
+  volume        = {666},
+  eid           = {A107},
+  pages         = {A107},
+  doi           = {10.1051/0004-6361/202244133},
+  archiveprefix = {arXiv},
+  eprint        = {2205.12962},
+  primaryclass  = {astro-ph.EP},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2022A&A...666A.107S},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{San2024,
+  doi       = {10.21105/joss.06648},
+  url       = {https://doi.org/10.21105/joss.06648},
+  year      = {2024},
+  publisher = {The Open Journal},
+  volume    = {9},
+  number    = {96},
+  pages     = {6648},
+  author    = {Metin San},
+  title     = {ZodiPy: A Python package for zodiacal light simulations},
+  journal   = {Journal of Open Source Software}
+} 
