+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "zodipy"
 version = "0.9.2"
-description = "Software for simulating zodiacal emission"
+description = "A Python package for zodiacal light simulations"
 authors = ["Metin San <metinisan@gmail.com>"]
 readme = "README.md"
 license = "GPL-3.0-or-later"
