### 2024-04-20 Cosmoglobe/zodipy#21 @pllim
Hello and I apologize for any inconvenience. Yes, synphot dropped support for Python 3.9 but only recently via https://github.com/spacetelescope/synphot_refactor/pull/387 . You should still be able to get `synphot < 1.4` in Python 3.9 .
