### 2024-04-20 Cosmoglobe/zodipy#21 @pllim
Would be happy to discuss more if you are interested. Integration is nothing too fancy. `synphot` uses trapezoid integration under the curve provided by scipy (used to be `np.trapz` in older versions).  See https://synphot.readthedocs.io/en/latest/api/synphot.spectrum.BaseSpectrum.html#synphot.spectrum.BaseSpectrum.integrate (analytical was added much later and defers to `astropy.modeling` but not widely used).
