### 2024-04-22 Cosmoglobe/zodipy#21 @MetinSa
Thanks it should work with `synphot < 1.4`.

Yeah that would be nice! I want to use `synphot` to compute a table of bandpass integrated blackbody emission values where the blackbodies have varying temperature. It should work for either a center frequency or an empirical bandpass provided by the user as two 1D arrays corresponding to weights and frequencies.  See `get_bandpass_interpolation_table` in https://github.com/Cosmoglobe/zodipy/blob/main/zodipy/_bandpass.py.
Not quite sure how I would do this using these models but `synphot.models.Empirical1d` or `synphot.spectrum.BaseSpectrum` seem relevant? I also see that there is a `synphot.models.Blackbody1D` which means that I probably don't need to use any custom `b_nu` function in ZodiPy either?
