### 2024-04-23 #161 @dhomeier
Congratulations again @MetinSa to your successful publication with pyOpenSci and JOSS! I'd like to take the opportunity to invite you and your maintainer team to write a blog post (totally optional) on your package for us to promote your work! If you are interested - here are a few examples of other blog posts:

[pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
[rdata](https://www.pyopensci.org/blog/read-r-datasets-from-python)
[SunPy](https://github.com/sunpy/sunpy.org/blob/main/posts/2024/2024-01-24-pyopensci.md) (external blog)

and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

It can even be a tutorial-like post that highlights what your package does and how to use it. Then we can share it with people to get the word out about your package.

If you are too busy for this, no worries. But if you have time - we'd love to spread the word about your package!
