### 2024-04-25 ce4ac11
Move solar_cut argument from the ZodiPy class to the binned function signatures, as these really only make sense in that context (I think). Also, start the migration from healpy to astropy-healpix
