### 2024-04-26 0bf63f8
Use angular separation instead of cone search. Turns out healpix-cone search is much slower than to just compute the angular separation between the pointing
