+++ tests/_strategies.py
@@ -198,7 +198,6 @@ MODEL_STRATEGY_MAPPINGS: dict[str, SearchStrategy[Any]] = {
     "model": sampled_from(AVAILABLE_MODELS),
     "gauss_quad_degree": integers(min_value=1, max_value=200),
     "extrapolate": booleans(),
-    "solar_cut": quantities(min_value=0, max_value=360, unit=u.deg),
 }
 
 
