### 2024-04-26 b484d74
Add tests for skycoord api
