+++ tests/test_get_emission.py
@@ -1,33 +1,55 @@
 from __future__ import annotations
 
+import astropy.coordinates as coords
 import astropy.units as u
 import healpy as hp
 import numpy as np
 import pytest
 from astropy.time import Time, TimeDelta
 from hypothesis import given, settings
-from hypothesis.strategies import DataObject, booleans, data, integers
+from hypothesis.strategies import DataObject, booleans, data
 
 from zodipy.zodipy import Zodipy
 
 from ._strategies import (
     angles,
-    freq,
-    model,
-    nside,
-    obs,
+    freqs,
+    nsides,
+    obs_positions,
     pixels,
     quantities,
-    random_freq,
-    time,
+    random_freqs,
+    sky_coords,
+    times,
     weights,
+    zodipy_models,
 )
 from ._tabulated_dirbe import DAYS, LAT, LON, TABULATED_DIRBE_EMISSION
 
 DIRBE_START_DAY = Time("1990-01-01")
 
 
-@given(model(), time(), nside(), data())
+@given(zodipy_models(), times(), sky_coords(), data())
+@settings(deadline=None)
+def test_get_emission_skycoord(
+    model: Zodipy,
+    time: Time,
+    coordinates: coords.SkyCoord,
+    data: DataObject,
+) -> None:
+    """Property test for get_emission_skycoord."""
+    observer = data.draw(obs_positions(model, time))
+    frequency = data.draw(freqs(model))
+    emission = model.get_emission_skycoord(
+        coordinates,
+        freq=frequency,
+        obs_time=time,
+        obs_pos=observer,
+    )
+    assert emission.size == 1
+
+
+@given(zodipy_models(), times(), nsides(), data())
 @settings(deadline=None)
 def test_get_emission_pix(
     model: Zodipy,
@@ -36,20 +58,20 @@ def test_get_emission_pix(
     data: DataObject,
 ) -> None:
     """Property test for get_emission_pix."""
-    observer = data.draw(obs(model, time))
-    frequency = data.draw(freq(model))
+    observer = data.draw(obs_positions(model, time))
+    frequency = data.draw(freqs(model))
     pix = data.draw(pixels(nside))
     emission = model.get_emission_pix(
-        frequency,
-        pixels=pix,
+        pix,
         nside=nside,
+        freq=frequency,
         obs_time=time,
-        obs=observer,
+        obs_pos=observer,
     )
     assert np.size(emission) == np.size(pix)
 
 
-@given(model(), time(), nside(), data())
+@given(zodipy_models(), times(), nsides(), data())
 @settings(deadline=None)
 def test_get_binned_emission_pix(
     model: Zodipy,
@@ -58,22 +80,22 @@ def test_get_binned_emission_pix(
     data: DataObject,
 ) -> None:
     """Property test for get_binned_emission_pix."""
-    observer = data.draw(obs(model, time))
-    frequency = data.draw(freq(model))
+    observer = data.draw(obs_positions(model, time))
+    frequency = data.draw(freqs(model))
     pix = data.draw(pixels(nside))
     cut_solar = data.draw(booleans())
     emission_binned = model.get_binned_emission_pix(
-        frequency,
-        pixels=pix,
+        pix,
         nside=nside,
+        freq=frequency,
         obs_time=time,
-        obs=observer,
+        obs_pos=observer,
         solar_cut=data.draw(quantities(20, 50, u.deg)) if cut_solar else None,
     )
     assert emission_binned.shape == (hp.nside2npix(nside),)
 
 
-@given(model(), time(), angles(), data())
+@given(zodipy_models(), times(), angles(), data())
 @settings(deadline=None)
 def test_get_emission_ang(
     model: Zodipy,
[…]
