+++ zodipy/_unit_vectors.py
@@ -1,29 +1,30 @@
-from __future__ import annotations
+# from __future__ import annotations
 
-from typing import TYPE_CHECKING, Sequence
+# from typing import TYPE_CHECKING, Sequence
 
-import healpy as hp
-import numpy as np
+# import healpy as hp
+# import numpy as np
 
-if TYPE_CHECKING:
-    import astropy.units as u
-    import numpy.typing as npt
-
-
-def get_unit_vectors_from_pixels(
-    coord_in: str, pixels: Sequence[int] | npt.NDArray[np.int64], nside: int
-) -> npt.NDArray[np.float64]:
-    """Return ecliptic unit vectors from HEALPix pixels representing some pointing."""
-    unit_vectors = np.asarray(hp.pix2vec(nside, pixels))
-    return np.asarray(hp.Rotator(coord=[coord_in, "E"])(unit_vectors))
+# if TYPE_CHECKING:
+#     import astropy.units as u
+#     import numpy.typing as npt
 
 
-def get_unit_vectors_from_ang(
-    coord_in: str,
-    phi: u.Quantity[u.rad] | u.Quantity[u.deg],
-    theta: u.Quantity[u.rad] | u.Quantity[u.deg],
-    lonlat: bool = False,
-) -> npt.NDArray[np.float64]:
-    """Return ecliptic unit vectors from sky angles representing some pointing."""
-    unit_vectors = np.asarray(hp.ang2vec(theta, phi, lonlat=lonlat)).transpose()
-    return np.asarray(hp.Rotator(coord=[coord_in, "E"])(unit_vectors))
+# def get_unit_vectors_from_pixels(
+#     coord_in: str, pixels: Sequence[int] | npt.NDArray[np.int64], nside: int
+# ) -> npt.NDArray[np.float64]:
+#     """Return ecliptic unit vectors from HEALPix pixels representing some pointing."""
+#     unit_vectors = np.asarray(hp.pix2vec(nside, pixels))
+#     return np.asarray(hp.Rotator(coord=[coord_in, "E"])(unit_vectors))
+
+
+# def get_unit_vectors_from_ang(
+#     coord_in: str,
+#     phi: u.Quantity[u.rad] | u.Quantity[u.deg],
+#     theta: u.Quantity[u.rad] | u.Quantity[u.deg],
+#     lonlat: bool = False,
+# ) -> npt.NDArray[np.float64]:
+#     """Return ecliptic unit vectors from sky angles representing some pointing."""
+#     unit_vectors = np.asarray(hp.ang2vec(theta, phi, lonlat=lonlat)).transpose()
+#     return np.asarray(hp.Rotator(coord=[coord_in, "E"])(unit_vectors))
+
