+++ zodipy/_types.py
@@ -8,3 +8,5 @@ Pixels = Union[int, Sequence[int], npt.NDArray[np.integer]]
 SkyAngles = Union[u.Quantity[u.deg], u.Quantity[u.rad]]
 FrequencyOrWavelength = Union[u.Quantity[u.Hz], u.Quantity[u.m]]
 ParameterDict = dict
+NumpyArray = Union[npt.NDArray[np.float64], npt.NDArray[np.int64]]
+MJySr = u.Quantity[u.MJy / u.sr]

[…]
