+++ zodipy/zodipy.py
@@ -3,31 +3,28 @@ from __future__ import annotations
 import functools
 import multiprocessing
 import platform
-from typing import TYPE_CHECKING, Callable, Sequence
+from typing import TYPE_CHECKING, Callable
 
-import astropy.coordinates as coords
-import astropy.units as u
 import astropy_healpix as hp
 import numpy as np
+from astropy import coordinates as coords
+from astropy import units
 from scipy import interpolate
 
 from zodipy._bandpass import get_bandpass_interpolation_table, validate_and_get_bandpass
 from zodipy._constants import SPECIFIC_INTENSITY_UNITS
+from zodipy._coords import get_earth_skycoord, get_obs_skycoord
 from zodipy._emission import EMISSION_MAPPING
 from zodipy._interpolate_source import SOURCE_PARAMS_MAPPING
 from zodipy._ipd_comps import ComponentLabel
 from zodipy._ipd_dens_funcs import construct_density_partials_comps
 from zodipy._line_of_sight import get_line_of_sight_start_and_stop_distances
-from zodipy._sky_coords import get_obs_and_earth_positions
 from zodipy._validators import get_validated_ang
 from zodipy.model_registry import model_registry
 
 if TYPE_CHECKING:
     import numpy.typing as npt
-    from astropy.time import Time
-
-    from zodipy._types import FrequencyOrWavelength, ParameterDict, Pixels, SkyAngles
-
+    from astropy import time
 
 PLATFORM = platform.system().lower()
 SYS_PROC_START_METHOD = "fork" if "windows" not in PLATFORM else None
@@ -90,11 +87,11 @@ class Zodipy:
         """Return a list of available observers given an ephemeris."""
         return [*list(coords.solar_system_ephemeris.bodies), "semb-l2"]
 
-    def get_parameters(self) -> ParameterDict:
+    def get_parameters(self) -> dict:
         """Return a dictionary containing the interplanetary dust model parameters."""
         return self._ipd_model.to_dict()
 
-    def update_parameters(self, parameters: ParameterDict) -> None:
+    def update_parameters(self, parameters: dict) -> None:
         """Update the interplanetary dust model parameters.
 
         Args:
@@ -120,12 +117,12 @@ class Zodipy:
         self,
         coord: coords.SkyCoord,
         *,
-        obs_time: Time,
-        freq: FrequencyOrWavelength,
-        obs_pos: u.Quantity[u.AU] | str = "earth",
-        weights: Sequence[float] | npt.NDArray[np.floating] | None = None,
+        obs_time: time.Time,
+        freq: units.Quantity,
+        obs_pos: units.Quantity | str = "earth",
+        weights: npt.ArrayLike | None = None,
         return_comps: bool = False,
-    ) -> u.Quantity[u.MJy / u.sr]:
+    ) -> units.Quantity[units.MJy / units.sr]:
         """Return the simulated zodiacal light for observations in an Astropy `SkyCoord` object.
 
         The pointing, for which to compute the emission, is specified in form of angles on
@@ -136,13 +133,13 @@ class Zodipy:
                 object must have a frame attribute representing the coordinate frame of the
                 input pointing, for example `astropy.coordinates.Galactic`. The frame must
                 be convertible to `BarycentricMeanEcliptic`.
-            obs_time: Time of observation.
+            obs_time: Time of observation. This should be a single observational time.
             freq: Delta frequency/wavelength or a sequence of frequencies corresponding to
                 a bandpass over which to evaluate the zodiacal emission. The frequencies
                 must be strictly increasing.
             obs_pos: The heliocentric ecliptic position of the observer in AU, or a string
                 representing an observer in the `astropy.coordinates.solar_system_ephemeris`.
-                Defaults to 'earth'.
[…]
