### 2024-04-27 f6ef2bc
Update pyproject stubs
