+++ pyproject.toml
@@ -33,6 +33,7 @@ healpy = "^1.16.6"
 astropy = ">=5.0.1"
 jplephem = "^2.17"
 scipy = "^1.13.0"
+astropy-healpix = "^1.0.3"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
@@ -58,8 +59,8 @@ build-backend = "poetry.core.masonry.api"
 disable_error_code = ["misc"]
 plugins = "numpy.typing.mypy_plugin"
 overrides = [
-    { module = "healpy.*", ignore_missing_imports = true },
     { module = "astropy.*", ignore_missing_imports = true },
+    { module = "astropy_healpix.*", ignore_missing_imports = true },
     { module = "scipy.*", ignore_missing_imports = true },
     { module = "pkg_resources.*", ignore_missing_imports = true },
 ]
