+++ .pre-commit-config.yaml
@@ -7,12 +7,16 @@ repos:
       langauge_version: python3
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    # Ruff version.
     rev: v0.3.7
     hooks:
       - id: ruff
       - id: ruff-format
 
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v1.10.0
+    hooks:
+    -   id: mypy
+        additional_dependencies: [numpy>=1.21.0]
 hooks:
   - id: poetry-export
     args: ["--dev", "-f", "requirements.txt", "-o", "requirements-dev.txt"]
