### 2024-04-27 824edbd
Revert some breaking API changes for the healpy compatible functions
