+++ zodipy/zodipy.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import functools
 import multiprocessing
 import platform
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING, Callable, Literal
 
 import astropy_healpix as hp
 import numpy as np
@@ -13,7 +13,7 @@ from scipy import interpolate
 
 from zodipy._bandpass import get_bandpass_interpolation_table, validate_and_get_bandpass
 from zodipy._constants import SPECIFIC_INTENSITY_UNITS
-from zodipy._coords import get_earth_skycoord, get_obs_skycoord
+from zodipy._coords import get_earth_skycoord, get_frame_from_string, get_obs_skycoord
 from zodipy._emission import EMISSION_MAPPING
 from zodipy._interpolate_source import SOURCE_PARAMS_MAPPING
 from zodipy._ipd_comps import ComponentLabel
@@ -117,31 +117,28 @@ class Zodipy:
         self,
         coord: coords.SkyCoord,
         *,
-        obs_time: time.Time,
         freq: units.Quantity,
-        obs_pos: units.Quantity | str = "earth",
         weights: npt.ArrayLike | None = None,
+        obs_pos: units.Quantity | str = "earth",
         return_comps: bool = False,
     ) -> units.Quantity[units.MJy / units.sr]:
-        """Return the simulated zodiacal light for observations in an Astropy `SkyCoord` object.
-
-        The pointing, for which to compute the emission, is specified in form of angles on
-        the sky given by `theta` and `phi`.
+        """Return the simulated zodiacal light for all observations in a `SkyCoord` object.
 
         Args:
-            coord: Astropy `SkyCoord` object representing the pointing on the sky. This
-                object must have a frame attribute representing the coordinate frame of the
-                input pointing, for example `astropy.coordinates.Galactic`. The frame must
-                be convertible to `BarycentricMeanEcliptic`.
-            obs_time: Time of observation. This should be a single observational time.
+            coord: `astropy.coordinates.SkyCoord` object representing the observations for which to
+                simulate the zodiacal light. The `frame` and `obstime` attributes of the `SkyCoord`
+                object must be set. The `obstime` attribute should correspond to a single
+                observational time for which the zodiacal light is assumed to be stationary.
+                Additionally, the frame must be convertible to
+                `astropy.coordinates.BarycentricMeanEcliptic`.
             freq: Delta frequency/wavelength or a sequence of frequencies corresponding to
                 a bandpass over which to evaluate the zodiacal emission. The frequencies
                 must be strictly increasing.
+            weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
+                are assumed to be given in spectral radiance units (Jy/sr).
             obs_pos: The heliocentric ecliptic position of the observer in AU, or a string
                 representing an observer in the `astropy.coordinates.solar_system_ephemeris`.
                 This should correspond to a single position. Defaults to 'earth'.
-            weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
-                are assumed to be given in spectral radiance units (Jy/sr).
             return_comps: If True, the emission is returned component-wise. Defaults to False.
 
         Returns:
@@ -153,13 +150,15 @@ class Zodipy:
             return_inverse=True,
             axis=1,
         )
-        coord = coords.SkyCoord(unique_lon * units.deg, unique_lat * units.deg, frame=coord.frame)
+        coord = coords.SkyCoord(
+            unique_lon * units.deg, unique_lat * units.deg, frame=coord.frame, obstime=coord.obstime
+        )
 
         return self._compute_emission(
[…]
