### 2024-04-27 2062d48
Remove healpy as a dependency. Add healpy as a dev dependency for plotting
