+++ zodipy/zodipy.py
@@ -18,7 +18,7 @@ from zodipy._emission import EMISSION_MAPPING
 from zodipy._interpolate_source import SOURCE_PARAMS_MAPPING
 from zodipy._ipd_comps import ComponentLabel
 from zodipy._ipd_dens_funcs import construct_density_partials_comps
-from zodipy._line_of_sight import get_line_of_sight_start_and_stop_distances
+from zodipy._line_of_sight import get_line_of_sight_range
 from zodipy._validators import get_validated_ang
 from zodipy.model_registry import model_registry
 
@@ -136,9 +136,9 @@ class Zodipy:
                 must be strictly increasing.
             weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
                 are assumed to be given in spectral radiance units (Jy/sr).
-            obs_pos: The heliocentric ecliptic position of the observer in AU, or a string
-                representing an observer in the `astropy.coordinates.solar_system_ephemeris`.
-                This should correspond to a single position. Defaults to 'earth'.
+            obs_pos: The heliocentric ecliptic position of the observer, or a string representing
+                an observer in the `astropy.coordinates.solar_system_ephemeris`. This should
+                correspond to a single position. Defaults to 'earth'.
             return_comps: If True, the emission is returned component-wise. Defaults to False.
 
         Returns:
@@ -151,7 +151,10 @@ class Zodipy:
             axis=1,
         )
         coord = coords.SkyCoord(
-            unique_lon * units.deg, unique_lat * units.deg, frame=coord.frame, obstime=coord.obstime
+            unique_lon * units.deg,
+            unique_lat * units.deg,
+            frame=coord.frame,
+            obstime=coord.obstime,
         )
 
         return self._compute_emission(
@@ -195,9 +198,9 @@ class Zodipy:
                 a bandpass over which to evaluate the zodiacal emission. The frequencies
                 must be strictly increasing.
             obs_time: Time of observation. This should be a single observational time.
-            obs_pos: The heliocentric ecliptic position of the observer in AU, or a string
-                representing an observer in the `astropy.coordinates.solar_system_ephemeris`.
-                This should correspond to a single position. Defaults to 'earth'.
+            obs_pos: The heliocentric ecliptic position of the observer, or a string representing
+                an observer in the `astropy.coordinates.solar_system_ephemeris`. This should
+                correspond to a single position. Defaults to 'earth'.
             coord_in: Coordinate frame of the input pointing. Assumes 'E' (ecliptic
                 coordinates) by default.
             weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
@@ -248,9 +251,9 @@ class Zodipy:
                 a bandpass over which to evaluate the zodiacal emission. The frequencies
                 must be strictly increasing.
             obs_time: Time of observation. This should be a single observational time.
-            obs_pos: The heliocentric ecliptic position of the observer in AU, or a string
-                representing an observer in the `astropy.coordinates.solar_system_ephemeris`.
-                This should correspond to a single position. Defaults to 'earth'.
+            obs_pos: The heliocentric ecliptic position of the observer, or a string representing
+                an observer in the `astropy.coordinates.solar_system_ephemeris`. This should
+                correspond to a single position. Defaults to 'earth'.
             coord_in: Coordinate frame of the input pointing. Assumes 'E' (ecliptic
                 coordinates) by default.
             weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
@@ -302,9 +305,9 @@ class Zodipy:
                 must be strictly increasing.
             weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
                 are assumed to be given in spectral radiance units (Jy/sr).
[…]
