+++ zodipy/_types.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import Sequence, Union
 
 import astropy.units as u
@@ -8,3 +10,4 @@ Pixels = Union[int, Sequence[int], npt.NDArray[np.integer]]
 SkyAngles = Union[u.Quantity[u.deg], u.Quantity[u.rad]]
 FrequencyOrWavelength = Union[u.Quantity[u.Hz], u.Quantity[u.m]]
 ParameterDict = dict
+Float = Union[float, np.floating]
