+++ docs/install.md
@@ -13,7 +13,7 @@ pip install zodipy
 ZodiPy has the following dependencies (these are automatically downloaded alongside ZodiPy):
 
 - [Astropy](https://www.astropy.org) (>= 5.0.1)
+- [Astropy-healpix](https://astropy-healpix.readthedocs.io/en/latest/)
 - [NumPy](https://numpy.org)
-- [healpy](https://healpy.readthedocs.io/en/latest/)
 - [jplehem](https://pypi.org/project/jplephem/)
 - [SciPy](https://scipy.org/)
