### 2024-04-28 b85fa59
Add tests to get back to 99% coverage
