+++ zodipy/_types.py
@@ -1,13 +1,0 @@
-from __future__ import annotations
-
-from typing import Sequence, Union
-
-import astropy.units as u
-import numpy as np
-import numpy.typing as npt
-
-Pixels = Union[int, Sequence[int], npt.NDArray[np.integer]]
-SkyAngles = Union[u.Quantity[u.deg], u.Quantity[u.rad]]
-FrequencyOrWavelength = Union[u.Quantity[u.Hz], u.Quantity[u.m]]
-ParameterDict = dict
-Float = Union[float, np.floating]
