+++ .pre-commit-config.yaml
@@ -16,6 +16,7 @@ repos:
     rev: v1.10.0
     hooks:
     -   id: mypy
+        exclude: tests/
         additional_dependencies: [numpy>=1.21.0]
 hooks:
   - id: poetry-export
