### 2024-04-28 Cosmoglobe/zodipy#21 @MetinSa
Thanks you, that helps!

I am still having some trouble... Is this the way to go if I want to bandpass integrate for a range of different blackbodies:
```python
bandpass = SpectralElement(Empirical1D, points=frequencies, lookup_table=weights)
for temp in temperatures:
    bb = SourceSpectrum(BlackBody1D(temp))
    obs = Observation(bb, bandpass)
    bp_integrated_emission = obs.integrate()
```
I am also interested in getting the emission in spectral flux density units, so for instance MJy/sr.
How efficient is this compared to just evaluation `b_nu(frequencies)` and then bandpass integrating with `np.trapz` for instance?
