### 2024-04-28 Cosmoglobe/zodipy#21 @pllim
@MetinSa , is that a real bandpass or you are using it to represent something else? If you have a minimal example with just numbers and expected answers, I can see what in `synphot` can best represent your use case. Thanks!
