### 2024-04-28 Cosmoglobe/zodipy#21 @MetinSa
@pllim yes the `frequency` and `weights` arrays represent a user inputted bandpass in normalized `Jy` units. I made a gist where i read in the DIRBE 25um bandpass and do the bandpass integration i wish to do in ZodiPy with synphot https://gist.github.com/MetinSa/9732a366fbe932a044819dcaa948df88.

Thank you for being so quick and nice to respond:)
