### 2024-04-29 d053282
Move bandpass calculations to Zodipy init fixing  #26 and fix bug with ephem
