+++ zodipy/zodipy.py
@@ -37,6 +37,11 @@ class Zodipy:
     either in sky angles or through HEALPix pixels.
 
     Args:
+        freq: Delta frequency/wavelength or a sequence of frequencies corresponding to
+            a bandpass over which to evaluate the zodiacal emission. The frequencies
+            must be strictly increasing.
+        weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
+            are assumed to be given in spectral radiance units (Jy/sr).
         model (str): Name of the interplanetary dust model to use in the simulations.
             Defaults to DIRBE.
         gauss_quad_degree (int): Order of the Gaussian-Legendre quadrature used to evaluate
@@ -60,27 +65,39 @@ class Zodipy:
 
     def __init__(
         self,
+        freq: units.Quantity,
+        weights: npt.ArrayLike | None = None,
         model: str = "dirbe",
         gauss_quad_degree: int = 50,
         extrapolate: bool = False,
         interp_kind: str = "linear",
-        ephemeris: str = "de432s",
+        ephemeris: str = "builtin",
         n_proc: int = 1,
     ) -> None:
         self.model = model
         self.gauss_quad_degree = gauss_quad_degree
         self.extrapolate = extrapolate
-        self.interp_kind = interp_kind
         self.ephemeris = ephemeris
         self.n_proc = n_proc
 
         self._interpolator = functools.partial(
             interpolate.interp1d,
-            kind=self.interp_kind,
+            kind=interp_kind,
             fill_value="extrapolate" if self.extrapolate else np.nan,
         )
         self._ipd_model = model_registry.get_model(model)
         self._gauss_points_and_weights = np.polynomial.legendre.leggauss(gauss_quad_degree)
+
+        bandpass = validate_and_get_bandpass(
+            freq=freq,
+            weights=weights,
+            model=self._ipd_model,
+            extrapolate=self.extrapolate,
+        )
+        self.bandpass_interpolatation_table = get_bandpass_interpolation_table(bandpass)
+        self.source_parameters = SOURCE_PARAMS_MAPPING[type(self._ipd_model)](
+            bandpass, self._ipd_model, self._interpolator
+        )
 
     @property
     def supported_observers(self) -> list[str]:
@@ -117,8 +134,6 @@ class Zodipy:
         self,
         coord: coords.SkyCoord,
         *,
-        freq: units.Quantity,
-        weights: npt.ArrayLike | None = None,
         obs_pos: units.Quantity | str = "earth",
         return_comps: bool = False,
     ) -> units.Quantity[units.MJy / units.sr]:
@@ -131,11 +146,6 @@ class Zodipy:
                 observational time for which the zodiacal light is assumed to be stationary.
                 Additionally, the frame must be convertible to
                 `astropy.coordinates.BarycentricMeanEcliptic`.
-            freq: Delta frequency/wavelength or a sequence of frequencies corresponding to
-                a bandpass over which to evaluate the zodiacal emission. The frequencies
-                must be strictly increasing.
-            weights: Bandpass weights corresponding the the frequencies in `freq`. The weights
-                are assumed to be given in spectral radiance units (Jy/sr).
             obs_pos: The heliocentric ecliptic position of the observer, or a string representing
                 an observer in the `astropy.coordinates.solar_system_ephemeris`. This should
                 correspond to a single position. Defaults to 'earth'.
@@ -158,8 +168,6 @@ class Zodipy:
         )
 
         return self._compute_emission(
-            freq=freq,
-            weights=weights,
             obs_pos=obs_pos,
             obs_time=obs_time,
             coordinates=coord,
@@ -173,11 +181,9 @@ class Zodipy:
         phi: units.Quantity,
         *,
         lonlat: bool = False,
-        freq: units.Quantity,
         obs_time: time.Time,
[…]
