### 2024-04-29 5d49729
Solar cut is now used to filter out pointing instead of post-evaluation masking
