+++ zodipy/zodipy.py
@@ -46,18 +46,18 @@ class Zodipy:
             Defaults to DIRBE.
         gauss_quad_degree (int): Order of the Gaussian-Legendre quadrature used to evaluate
             the line-of-sight integral in the simulations. Default is 50 points.
-        interp_kind (str): Interpolation kind used to interpolate relevant model parameters.
-            Defaults to 'linear'. For more information on available interpolation methods,
-            please visit the [Scipy documentation](
-            https://docs.scipy.org/doc/scipy/reference/generated/scipy.interpolate.interp1d.html).
+        interp_kind (str): Interpolation kind used in `scipy.interpolate.interp1d` to interpolate
+            spectral paramters (see [Scipy documentation](
+            https://docs.scipy.org/doc/scipy/reference/generated/scipy.interpolate.interp1d.html)).
+            Defaults to 'linear'.
         extrapolate (bool): If `True` all spectral quantities in the selected model are
             extrapolated to the requested frequencies or wavelengths. If `False`, an
             exception is raised on requested frequencies/wavelengths outside of the
             valid model range. Default is `False`.
-        ephemeris (str): Ephemeris used to compute the positions of the observer and the
-            Earth. Defaults to 'de432s', which requires downloading (and caching) a ~10MB
-            file. For more information on available ephemeridis, please visit the [Astropy
-            documentation](https://docs.astropy.org/en/stable/coordinates/solarsystem.html)
+        ephemeris (str): Ephemeris used in `astropy.coordinates.solar_system_ephemeris` to compute
+            the positions of the observer and the Earth. Defaults to 'builtin'. See the
+            [Astropy documentation](https://docs.astropy.org/en/stable/coordinates/solarsystem.html)
+            for available ephemerides.
         n_proc (int): Number of cores to use. If `n_proc` is greater than 1, the line-of-sight
             integrals are parallelized using the `multiprocessing` module. Defaults to 1.
 
@@ -99,37 +99,6 @@ class Zodipy:
             bandpass, self._ipd_model, self._interpolator
         )
 
-    @property
-    def supported_observers(self) -> list[str]:
-        """Return a list of available observers given an ephemeris."""
-        return [*list(coords.solar_system_ephemeris.bodies), "semb-l2"]
-
-    def get_parameters(self) -> dict:
-        """Return a dictionary containing the interplanetary dust model parameters."""
-        return self._ipd_model.to_dict()
-
-    def update_parameters(self, parameters: dict) -> None:
-        """Update the interplanetary dust model parameters.
-
-        Args:
-            parameters: Dictionary of parameters to update. The keys must be the names
-                of the parameters as defined in the model. To get the parameters dict
-                of an existing model, use `Zodipy("dirbe").get_parameters()`.
-
-        """
-        _dict = parameters.copy()
-        _dict["comps"] = {}
-        for key, value in parameters.items():
-            if key == "comps":
-                for comp_key, comp_value in value.items():
-                    _dict["comps"][ComponentLabel(comp_key)] = type(
-                        self._ipd_model.comps[ComponentLabel(comp_key)]
-                    )(**comp_value)
-            elif isinstance(value, dict):
-                _dict[key] = {ComponentLabel(k): v for k, v in value.items()}
-
-        self._ipd_model = self._ipd_model.__class__(**_dict)
-
     def get_emission_skycoord(
         self,
         coord: coords.SkyCoord,
@@ -160,10 +129,16 @@ class Zodipy:
             return_inverse=True,
             axis=1,
         )
+
         obs_time = coord.obstime
+        if obs_time is None:
+            msg = "The `obstime` attribute of the `SkyCoord` object must be set."
[…]
