### 2024-04-29 090b557
Attempts to fix #27, currently fails due to keyword arguments in interp_kind.
