### 2024-04-29 333b3f2
Move frequency and weights to Zodipy initializer and away from methods
