### 2024-04-29 57056e2
Fix bug with freqs hypothesis strategy
