### 2024-04-29 Cosmoglobe/zodipy#21 @pllim
When I run your code, I find that `normalized_weights` (the actual numbers multiplied to blackbody flux) is all negative. In the context of "bandpass", in which the values should only be between 0 and 1 (inclusive), this is un-physical. Could you please clarify?

![Screenshot 2024-04-29 140856](https://github.com/Cosmoglobe/zodipy/assets/2090236/3defefd0-1b2c-46e4-b355-ab230a465818)
