### 2024-04-29 Cosmoglobe/zodipy#21 @MetinSa
Ahh, I had forgotton to flip the `frequencies` and `weights` arrays after changing to frequency convention.
