### 2024-04-29 Cosmoglobe/zodipy#21 @pllim
Just for my own record, looks like you are only interested in this part of the blackbody curve.

![Screenshot 2024-04-29 142956](https://github.com/Cosmoglobe/zodipy/assets/2090236/259d9477-9992-4733-94d6-624cf143ba9c)
