### 2024-04-29 Cosmoglobe/zodipy#21 @MetinSa
Right, for this particular band pass, but it should be able to handle arbitrary frequency/wavelength ranges and temperatures between 50-300K
