### 2024-04-29 79aa169
Use built in `astropy.models.physical_models.BlackBody`
