+++ zodipy/_constants.py
@@ -1,10 +1,5 @@
-import astropy.constants as const
 import astropy.units as u
 import numpy as np
-
-h: float = const.h.value
-c: float = const.c.value
-k_B: float = const.k_B.value
 
 SPECIFIC_INTENSITY_UNITS = u.W / u.Hz / u.m**2 / u.sr
 
