### 2024-04-29 Cosmoglobe/zodipy#21 @pllim
I don't grok the bandpass normalization, but that doesn't matter.

I tried re-constructing what I think you are doing with `synphot` API but I cannot get the numbers to match exactly. If all you want to do is simple trapezoid integration with some custom normalization, `synphot` API might be a little of an overkill, because there are a lot of assumptions and checks that went into full-blown synthetic photometry. Furthermore, for historical reason, all the internal calculations are done in wavelength space but looks like you want to calculate in frequency space natively.

However, `synphot` does have a [synphot.blackbody.blackbody_nu](https://synphot.readthedocs.io/en/latest/api/synphot.blackbody.blackbody_nu.html#synphot.blackbody.blackbody_nu) function that you can use to replace your `b_nu`. p.s. I noticed that your `b_nu` is missing the per-steradian in returned flux unit.

Also, in tracing your calculation, I think you need to get rid of the per-micron in the integrated flux unit (because you already integrated over a range of Hz/micron), so the answer I think should be `2.2052944808204042 W / m2` and not `2.2052944808204042 W / (um m2)`.

I hope this helps and thanks for considering `synphot`.
