### 2024-04-29 Cosmoglobe/zodipy#21 @pllim
Also, in numpy 2.0, `np.trapz` has been deprecated in favor of `scipy.integrate.trapezoid`. FYI.
