### 2024-04-30 Cosmoglobe/zodipy#21 @MetinSa
I had the feeling that `synphot` might be a bit overkill for this purpose yes, but this was useful for me nevertheless. I will simply move over to use `astropy.models.physical_model.BlackBody` which can handle wavelengths too , instead of the custom `b_nu` implementation.

> Also, in numpy 2.0, np.trapz has been deprecated in favor of scipy.integrate.trapezoid. FYI.

Thats a good catch, thanks @pllim!
