### 2024-04-30 Cosmoglobe/zodipy#21 @pllim
Yes, the blackbody model might work too. Good luck!
