### 2024-05-01 37366c9
Refactor bandpass implementation
