+++ zodipy/source_params.py
@@ -220,10 +220,41 @@ ALBEDO_DIRBE = {
     ),
 }
 
-PHASE_FUNCTION_DIRBE = [
-    (-0.94209999, -0.52670002, -0.4312, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0),
-    (0.1214, 0.18719999, 0.1715, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0),
-    (-0.1648, -0.59829998, -0.63330001, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0),
+C1_DIRBE = [
+    -0.94209999,
+    -0.52670002,
+    -0.4312,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+]
+C2_DIRBE = [
+    0.1214,
+    0.18719999,
+    0.1715,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+]
+C3_DIRBE = [
+    -0.1648,
+    -0.59829998,
+    -0.63330001,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
+    0.0,
 ]
 
 SOLAR_IRRADIANCE_DIRBE = [
