+++ tests/test_reprs.py
@@ -9,4 +9,4 @@ from ._strategies import zodipy_models
 def test_ipd_model_repr(model: Zodipy) -> None:
     """Tests that the IPD model has a userfriendly repr."""
     repr(model)
-    repr(model._ipd_model)
+    repr(model._interplanetary_dust_model)

[…]
