### 2024-05-01 Cosmoglobe/zodipy#20 @MetinSa
This is now done internally with the `SkyCoord` object in `v1.0.0`
