### 2024-05-01 Cosmoglobe/zodipy#19 @MetinSa
This exists in `v1.0.0`.
