### 2024-05-01 86114d2
Rename `Zodipy` to `Model` and `get_emission_skycoord` to `evaluate`. Remove all other methods.
