+++ zodipy/__init__.py
@@ -1,10 +1,10 @@
 from zodipy import comps, source_params
 from zodipy._contour import tabulate_density
 from zodipy.model_registry import model_registry
-from zodipy.zodipy import Zodipy
+from zodipy.zodipy import Model
 
 __all__ = (
-    "Zodipy",
+    "Model",
     "model_registry",
     "comps",
     "source_params",
