+++ docs/examples/get_binned_emission_solar_cutoff.py
@@ -1,31 +1,0 @@
-from multiprocessing import cpu_count
-
-import astropy.units as u
-import healpy as hp
-import matplotlib.pyplot as plt
-import numpy as np
-from astropy.time import Time
-
-from zodipy import Zodipy
-
-model = Zodipy("dirbe", n_proc=cpu_count())
-nside = 256
-
-binned_emission = model.get_binned_emission_pix(
-    np.arange(hp.nside2npix(nside)),
-    nside=nside,
-    freq=25 * u.micron,
-    obs_time=Time("2020-01-01"),
-    obs_pos="earth",
-    solar_cut=60 * u.deg,
-)
-
-hp.mollview(
-    binned_emission,
-    title="Solar cutoff at 60 degrees",
-    unit="MJy/sr",
-    max=80,
-    coord="E",
-    cmap="afmhot",
-)
-plt.show()
