+++ tests/dirbe_tabulated.py
@@ -1,3 +1,6 @@
+from astropy import time, units
+
+DIRBE_START_DAY = time.Time("1990-01-01")
 # Days from 1990-01-01 where 1 = 1 Jan, 1990
 DAYS = [162, 61, 244, 163, 225, 354, 144, 360, 289, 204]
 
@@ -151,3 +154,175 @@ TABULATED_DIRBE_EMISSION = {
         0.24349573391915608,
     ],
 }
+
+DIRBE_25um_BANDPASS: tuple[units.Quantity, list[float]] = (
+    [
+        15.543,
+        15.664,
+        15.787,
+        15.910,
+        16.034,
+        16.160,
+        16.286,
+        16.413,
+        16.541,
+        16.670,
+        16.801,
+        16.932,
+        17.064,
+        17.198,
+        17.332,
+        17.467,
+        17.604,
+        17.741,
+        17.880,
+        18.019,
+        18.160,
+        18.302,
+        18.445,
+        18.589,
+        18.734,
+        18.881,
+        19.028,
+        19.177,
+        19.327,
+        19.477,
+        19.630,
+        19.783,
+        19.938,
+        20.093,
+        20.250,
+        20.408,
+        20.568,
+        20.728,
+        20.890,
+        21.054,
+        21.218,
+        21.384,
+        21.551,
+        21.719,
+        21.889,
+        22.060,
+        22.232,
+        22.406,
+        22.581,
+        22.757,
+        22.935,
+        23.114,
+        23.295,
+        23.477,
+        23.660,
+        23.845,
+        24.031,
+        24.219,
+        24.408,
+        24.599,
+        24.791,
+        24.984,
+        25.180,
+        25.376,
+        25.574,
+        25.774,
+        25.976,
+        26.178,
+        26.383,
+        26.589,
+        26.797,
+        27.006,
+        27.217,
+        27.430,
+        27.644,
+        27.860,
+        28.077,
+        28.297,
+        28.518,
+        28.741,
+        28.965,
+        29.191,
+    ]
+    * units.micron,
+    [
+        0.01,
+        0.02,
+        0.04,
+        0.05,
+        0.08,
+        0.16,
+        0.21,
+        0.34,
+        0.45,
+        0.55,
+        0.67,
+        0.79,
+        0.85,
+        0.89,
+        0.92,
+        0.93,
+        0.91,
+        0.88,
+        0.84,
+        0.82,
+        0.88,
+        0.94,
+        1.00,
+        0.99,
+        0.93,
+        0.87,
+        0.81,
+        0.68,
+        0.56,
+        0.46,
+        0.51,
+        0.60,
+        0.69,
+        0.70,
+        0.64,
+        0.60,
+        0.58,
+        0.63,
+        0.66,
+        0.70,
+        0.75,
+        0.80,
+        0.80,
+        0.69,
+        0.59,
+        0.53,
+        0.57,
+        0.61,
+        0.66,
+        0.74,
+        0.83,
+        0.83,
+        0.78,
+        0.73,
+        0.70,
+        0.68,
+        0.67,
+        0.79,
+        0.90,
+        0.95,
+        0.94,
+        0.93,
+        0.77,
+        0.62,
+        0.48,
+        0.35,
+        0.24,
+        0.18,
+        0.14,
+        0.09,
+        0.08,
+        0.07,
+        0.08,
+        0.08,
+        0.07,
+        0.05,
+        0.04,
+        0.03,
+        0.02,
+        0.02,
+        0.01,
+        0.01,
+    ],
+)
