+++ zodipy/zodipy.py
@@ -3,21 +3,20 @@ from __future__ import annotations
 import functools
 import multiprocessing
 import platform
-from typing import TYPE_CHECKING, Callable
+from typing import TYPE_CHECKING
 
 import numpy as np
 from astropy import coordinates as coords
 from astropy import units
 from scipy import integrate
 
-from zodipy._coords import get_earth_skycoord, get_obs_skycoord
-from zodipy._emission import EMISSION_MAPPING
-from zodipy._ipd_comps import ComponentLabel
-from zodipy._ipd_dens_funcs import construct_density_partials_comps
-from zodipy._line_of_sight import get_line_of_sight_range
 from zodipy.blackbody import tabulate_bandpass_integrated_bnu, tabulate_center_wavelength_bnu
 from zodipy.interpolate import get_model_to_dicts_callable
+from zodipy.line_of_sight import get_line_of_sight_range, integrate_gauss_legendre
 from zodipy.model_registry import model_registry
+from zodipy.number_density import construct_density_partials_comps
+from zodipy.skycoords import get_earth_skycoord, get_obs_skycoord
+from zodipy.zodiacal_component import ComponentLabel
 
 if TYPE_CHECKING:
     import numpy.typing as npt
@@ -27,13 +26,7 @@ SYS_PROC_START_METHOD = "fork" if "windows" not in PLATFORM else None
 
 
 class Model:
-    """Main interface to ZodiPy.
-
-    The zodiacal light simulations are configured by specifying a bandpass (`freq`, `weights`)
-    or a delta/center frequency (`freq`), and a string representation of a built in interplanetary
-    dust model (`model`). See https://cosmoglobe.github.io/zodipy/introduction/ for a list of
-    available models.
-    """
+    """Main interface to ZodiPy."""
 
     def __init__(
         self,
@@ -53,7 +46,8 @@ class Model:
                 `weights` must be provided.
             weights: Bandpass weights corresponding the the frequencies in `freq`. The weights are
                 assumed to be given in spectral radiance units (Jy/sr).
-            name: Interplanetary dust model to use. Defaults to 'dirbe' which is the DIRBE model.
+            name: Interplanetary dust model to use. For a list of available models, see
+                https://cosmoglobe.github.io/zodipy/introduction/. Defaults to 'dirbe'.
             gauss_quad_degree: Order of the Gaussian-Legendre quadrature used to evaluate the
                 line-of-sight integral in the simulations. Default is 50 points.
             extrapolate: If `True` all spectral quantities in the selected model are extrapolated to
@@ -75,6 +69,7 @@ class Model:
             raise ValueError(msg)
 
         self._interplanetary_dust_model = model_registry.get_model(name)
+
         if not extrapolate and not self._interplanetary_dust_model.is_valid_at(x):
             msg = (
                 "The requested frequencies are outside the valid range of the model."
@@ -152,11 +147,13 @@ class Model:
         )
 
         skycoord = skycoord.transform_to(coords.BarycentricMeanEcliptic)
+        unit_vector = skycoord.cartesian.xyz.value
+        obspos = obs_skycoord.cartesian.xyz.to_value(units.AU)
 
         start, stop = get_line_of_sight_range(
             components=self._interplanetary_dust_model.comps.keys(),
-            unit_vectors=skycoord.cartesian.xyz.value,
-            obs_pos=obs_skycoord.cartesian.xyz.to_value(units.AU),
+            unit_vectors=unit_vector,
+            obs_pos=obspos,
         )
 
         density_partials = construct_density_partials_comps(
@@ -167,17 +164,15 @@ class Model:
                 ]
             },
         )
-
         common_integrand = functools.partial(
-            EMISSION_MAPPING[type(self._interplanetary_dust_model)],
[…]
