### 2024-05-02 75c7696
Improve performance by avoiding copying the skycoord object and instead using indexing to filter it
