+++ README.md
@@ -26,23 +26,26 @@ See the [documentation](https://cosmoglobe.github.io/zodipy/) for more informati
 ## A simple example
 ```python
 import astropy.units as u
+from astropy.coordinates import SkyCoord
 from astropy.time import Time
 
-from zodipy import Zodipy
+from zodipy import Model
 
+model = Model(25*u.micron)
 
-model = Zodipy(model="dirbe")
-
-emission = model.get_emission_ang(
-    25 * u.micron,
-    theta=[10, 10.1, 10.2] * u.deg,
-    phi=[90, 89, 88] * u.deg,
-    obs_time=Time("2022-01-01 12:00:00"),
-    obs="earth",
+lon = [10, 10.1, 10.2] * u.deg
+lat = [90, 89, 88] * u.deg
+skycoord = SkyCoord(
+    lon,
+    lat,
+    obstime=Time("2022-01-01 12:00:00"),
+    frame="galactic",
 )
 
+emission = model.evaluate(skycoord, obspos="earth")
+
 print(emission)
-#> [15.35392831 15.35495051 15.35616009] MJy / sr
+#> [27.52410841 27.66581351 27.81270207] MJy / sr
 ```
 
 ## Related scientific papers
