+++ zodipy/_constants.py
@@ -1,20 +1,0 @@
-import astropy.units as u
-import numpy as np
-
-SPECIFIC_INTENSITY_UNITS = u.W / u.Hz / u.m**2 / u.sr
-
-R_0 = np.finfo(np.float64).eps
-R_MARS = 1.5237
-R_VERITAS = 2.16
-R_EOS = 3.02
-R_THEMIS = 3.14
-R_EARTH = 1
-R_ASTEROID_BELT = 3.137
-R_NARROW_BANDS = 3.015
-R_JUPITER = 5.2
-R_KUIPER_BELT = 30
-DISTANCE_FROM_EARTH_TO_SEMB_L2 = u.Quantity(0.009896235034000056, u.AU)
-
-N_INTERPOLATION_POINTS = 100
-MIN_INTERPOLATION_GRID_TEMPERATURE = 40
-MAX_INTERPOLATION_GRID_TEMPERATURE = 550

[…]
