### 2024-05-06 3efb6bb
Move astropy_healpix to dev dependancy. Update dev dependencies. Finish writing docs
