+++ docs/usage.md
@@ -1,19 +1,20 @@
+!!! warning "Breaking API changes in `v1.0.0` and newer"
+    From version `v1.0.0` and out ZodiPy no implements the `get_emission_ang` and `get_emission_pix` methods. See the section on [HEALPix maps](usage.md#healpix-maps) for an example of how to make simulations from HEALPix pixel indices. We apologize for the inconvenience.  
+
 As an Astropy-affiliated package, ZodiPy is highly integrated with the astropy ecosystem.
 To make zodiacal light simulations, the `astropy.units`, `astropy.coordinates`, and `astropy.time` modules are used to provide user input. The coordinates for which ZodiPy will simulate the zodiacal light is specified in through the `astropy.coordinates.SkyCoord` object. Using ZodiPy is very simple and the user will only interact with *one* object `zodipy.Model`, which has *one* method `evaluate`.
 
-!!! danger "Breaking API changes in `v1.0.0`"
-    ZodiPy version >=`v1.0.0` no longer supports simulating zodiacal light from non `SkyCoord` coordinate inputs.
 ## Initializing Zodipy
 ### Working with the `Model` class
 The interface to ZodiPy is the `zodipy.Model` class
-```python
+```py hl_lines="2 4"
 import astropy.units as u
 import zodipy
 
 model = zodipy.Model(25 * u.micron)
 ```
-It has *one* required positional argument `x`, which represents a center wavelength/frequency or the points of a empirical bandpass. If `x` represents the points of a bandpass, the `weights` argument must also be provided
-```python 
+It has *one* required positional argument `x`, which represents a center wavelength/frequency or the points of an empirical bandpass. If `x` represents the points of a bandpass, the `weights` argument must also be provided
+```py hl_lines="4 5 7"
 import astropy.units as u
 import zodipy
 
@@ -24,7 +25,7 @@ model = zodipy.Model(points, weights=weights)
 ```
 
 ZodiPy supports several zodiacal light models (see the [introduction](introduction.md) page for more information regarding the supported models), which are all valid in wavelength/frequency ranges. By default, the `Model` object will initialize using the DIRBE model. To select a different model, we specify the keyword argument `name`
-```python
+```py hl_lines="4"
 import astropy.units as u
 import zodipy
 
@@ -33,35 +34,37 @@ model = zodipy.Model(25 * u.micron, name="planck18")
 
 ### Multiprocessing
 ZodiPy will distribute the input coordinates to cores if the keyword argument `n_proc` is `>= 1` using Python's `multiprocessing` module.
-```python
-import multiprocessing
 
-import astropy.units as u
-import zodipy
+=== "Specifying number of processes"
+    ```py hl_lines="4"
+    import astropy.units as u
+    import zodipy
 
-model = zodipy.Model(25 * u.micron, n_proc=multiprocessing.cpu_count())
-```
-Alternatively, a custom pool (`multiprocessing.pool.Pool`) may be provided.
+    model = zodipy.Model(25 * u.micron, n_proc=4)
+    ```
+=== "Custom pool object"
 
-```python
-import multiprocessing
+    ```py hl_lines="1 6 7"
+    import multiprocessing
 
-import astropy.units as u
-import zodipy
+    import astropy.units as u
+    import zodipy
 
-pool = multiprocessing.Pool(multiprocesing.cpu_count())
-model = zodipy.Model(25 * u.micron, pool=pool)
-```
+    pool = multiprocessing.Pool(4)
+    model = zodipy.Model(25 * u.micron, pool=pool)
+    ```
+
 !!! tip 
-    For all available optional keyword arguments in `zodipy.Model` see [the API reference](reference.md).
+    For all available optional keyword arguments in `zodipy.Model` see [the API reference](reference.md#zodipy.zodipy.Model).
 
 ## Simulating zodiacal light
[…]
