+++ .pre-commit-config.yaml
@@ -2,9 +2,9 @@ repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.3.0
     hooks:
-    - id: check-added-large-files
-    - id: check-yaml
-      langauge_version: python3
+      - id: check-added-large-files
+      - id: check-yaml
+        langauge_version: python3
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.3.7
@@ -15,9 +15,31 @@ repos:
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.10.0
     hooks:
-    -   id: mypy
-        exclude: '(tests|docs)/.*'
+      - id: mypy
+        exclude: "(tests|docs)/.*"
         additional_dependencies: [numpy>=1.21.0]
-hooks:
-  - id: poetry-export
-    args: ["--dev", "-f", "requirements.txt", "-o", "requirements-dev.txt"]
+
+  - repo: local
+    hooks:
+      - id: poetry-check
+        name: poetry-check
+        description: run poetry check to validate config
+        entry: poetry check
+        language: python
+        pass_filenames: false
+        files: ^(.*/)?pyproject\.toml$
+      - id: poetry-lock
+        name: poetry-lock
+        description: run poetry lock to update lock file
+        entry: poetry lock
+        language: python
+        pass_filenames: false
+        files: ^(.*/)?(poetry\.lock|pyproject\.toml)$
+      - id: poetry-export
+        name: poetry-export
+        description: run poetry export to sync lock file with requirements.txt
+        entry: poetry export
+        language: python
+        pass_filenames: false
+        files: ^(.*/)?poetry\.lock$
+        args: ["--dev", "-f", "requirements.txt", "-o", "requirements-dev.txt"]
