### 2024-05-07 622501b
Update pyproject.toml and docs
