+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "zodipy"
-version = "0.9.2"
+version = "1.0.0"
 description = "A Python package for zodiacal light simulations"
 authors = ["Metin San <metinisan@gmail.com>"]
 readme = "README.md"
@@ -102,6 +102,7 @@ lint.select = [
     "PLW",
     "TRY",
     "RUF",
+    "NPY201"
 ]
 line-length = 100
 lint.ignore = [
