+++ requirements.txt
@@ -1,0 +1,165 @@
+astropy-iers-data==0.2024.5.6.0.29.28 ; python_version >= "3.9" \
+    --hash=sha256:7f02a20d4bc72c22533123734cc48d694cdd32adc471c5d6659218d788b611da \
+    --hash=sha256:ad992722705af68b4fe3b79983c05c4bbed01b3ed31466fe6f574ddfb986b620
+astropy==6.0.1 ; python_version >= "3.9" \
+    --hash=sha256:034dff5994428fb89813f40a18600dd8804128c52edf3d1baa8936eca3738de4 \
+    --hash=sha256:129ed1fb1d23e6fbf8b8e697c2e7340d99bc6271b8c59f9572f3f47063a42e6a \
+    --hash=sha256:15a5da8a0a84d75b55fafd56630578131c3c9186e4e486b4d2fb15c349b844d0 \
+    --hash=sha256:1db9e95438472f6ed53fa2f4e2811c2d84f4085eeacc3cb8820d770d1ea61d1c \
+    --hash=sha256:1f183ab42655ad09b064a4e8eb9cd1eaa138b90ca2f0cd82a200afda062063a5 \
+    --hash=sha256:242b8f101301ab303366109d0dfe3cf0db745bf778f7b859fb486105197577d1 \
+    --hash=sha256:2b5ff962b0e586953f95b63ec047e1d7a3b6a12a13d11c6e909e0bcd3e05b445 \
+    --hash=sha256:2c00922548a666b026e2630a563090341d74c8222066e9c84c9673395bca7363 \
+    --hash=sha256:2f53caf9efebcc9040a92c977dcdae78dd0ff4de218fd316e4fcaffd9ace8dc1 \
+    --hash=sha256:34fd2bb39cbfa6a8815b5cc99008d59057b9d341db00c67dbb40a3784a8dfb08 \
+    --hash=sha256:46cbadf360bbadb6a106217e104b91f85dd618658caffdaab5d54a14d0d52563 \
+    --hash=sha256:4fdd54fa57b85d50c4b83ab7ffd90ba2ffcc3d725e3f8d5ffa1ff5f500ef6b97 \
+    --hash=sha256:5208b6f10956ca92efb73375364c81a7df365b441b07f4941a24ee0f1bd9e292 \
+    --hash=sha256:6e998ee0ffa58342b4d44f2843b036015e3a6326b53185c5361fea4430658466 \
+    --hash=sha256:6f28facb5800c0617f233c1db0e622da83de1f74ca28d0ff8646e360d4fda74e \
+    --hash=sha256:764992af1ee1cd6d6f26373d09ddb5ede639d025ce9ff658b3b6580dc2ba4ec6 \
+    --hash=sha256:87ebbae7ba52f4de9b9f45029a3167d6515399138048d0b734c9033fda7fd723 \
+    --hash=sha256:89a975de356d0608e74f1f493442fb3acbbb7a85b739e074460bb0340014b39c \
+    --hash=sha256:8bc267738a85f633142c246dceefa722b653e7ba99f02e86dd9a7b980467eafc \
+    --hash=sha256:8fbd6d88935749ae892445691ac0dbd1923fc6d8094753a35150fc7756118fe3 \
+    --hash=sha256:9b3bf27c51fb46bba993695eebd0c39a4e2a792b707e65b28ac4e8ae703f93d4 \
+    --hash=sha256:c33e3d746c3e7a324dbd76b236fe1e44304d5b6d941a1f724f419d01666d6d88 \
+    --hash=sha256:c682967736228cc4477e63db0e8854375dd31d755de55b30256de98f1f7b7c23 \
[…]
