### 2024-05-13 #161 @dhomeier
Closing the issue as completed as ZodiPy is now listed on https://www.pyopensci.org/python-packages.html

Feel free to follow up with a blog post whenever convenient; also if you wish to join the pyOpenSci Slack space @dncnwtts or @hermda02, you are welcome to send us an email address to be signed up with.
